Moved and changed the authentication task to sites webauth extras task
